maitake_sync
Module rwlock

An asynchronous readers-writer lock.

See the documentation for the RwLock type for details.

Structs

OwnedRwLockReadGuard

Owned RAII structure used to release the shared read access of a RwLock when dropped.

OwnedRwLockWriteGuard

Owned RAII structure used to release the exclusive write access of a RwLock when dropped.

RwLock

An asynchronous readers-writer lock.

RwLockReadGuard

RAII structure used to release the shared read access of a RwLock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a RwLock when dropped.