maitake_sync

Module wait_cell

An atomically registered Waker, for waking a single task.

See the documentation for the WaitCell type for details.

Structs

Subscribe

Future returned from WaitCell::subscribe().

Wait

Future returned from WaitCell::wait().

WaitCell

An atomically registered Waker.

Enums

PollWaitError

An error indicating that a WaitCell was closed or busy while attempting register a Waker.