maitake_sync
Module wait_queue

A queue of waiting tasks that can be woken in first-in, first-out order (or all at once).

See the WaitQueue type’s documentation for details.

Structs

Wait

Future returned from WaitQueue::wait().

WaitOwned

Future returned from WaitQueue::wait_owned().

WaitQueue

A queue of waiting tasks which can be woken in first-in, first-out order, or all at once.